+++ src/main/java/ru/golfstream/project/repos/TypeEmployeeRepo.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.golfstream.project.entity.TypeEmployee;
 @Repository
-public interface TypeEmployeeRepo extends JpaRepository<TypeEmployee, Integer> {
+public interface TypeEmployeeRepo extends JpaRepository<TypeEmployee, Long> {
 }
