+++ src/main/java/ru/golfstream/project/rest/dto/PurchaseDto.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 @Data
 @Builder
 public class PurchaseDto {
-    private Integer idVoucher;
-    private Integer idClient;
+    private Long idVoucher;
+    private Long idClient;
     private LocalDate dateOfPurchase;
 }
